refactor: move resolveRoutePolicyGUID to route_policy.go
The Get/UpdateRoutePolicyLabels wrappers in label.go follow the same
convention as every other resource (GetRouteLabels, GetDomainLabels,
etc.): the thin label wrapper stays in label.go and delegates GUID
resolution to a helper in the resource's own file. Move the
route-policy resolver alongside the other route-policy actor methods
and drop the now-unused ccv3 import from label.go.
