+++ actor/v7action/route_policy.go
@@ -146,6 +146,41 @@ func (actor Actor) DeleteRoutePolicyBySource(domainName, source, hostname, path 
 	return allWarnings, err
 }
 
+// resolveRoutePolicyGUID finds the GUID of the route policy to operate on.
+// If source is non-empty, it finds the policy with that source.
+// If source is empty, it requires exactly one policy (returns ambiguity or not-found error).
+func (actor *Actor) resolveRoutePolicyGUID(routeURL, spaceGUID, source string) (string, *resources.Metadata, Warnings, error) {
+	route, routeWarnings, err := actor.GetRoute(routeURL, spaceGUID)
+	if err != nil {
+		return "", nil, routeWarnings, err
+	}
+
+	routePolicies, _, policyWarnings, err := actor.CloudControllerClient.GetRoutePolicies(
+		ccv3.Query{Key: ccv3.RouteGUIDFilter, Values: []string{route.GUID}},
+	)
+	allWarnings := append(routeWarnings, Warnings(policyWarnings)...)
+	if err != nil {
+		return "", nil, allWarnings, err
+	}
+
+	if source != "" {
+		for _, p := range routePolicies {
+			if p.Source == source {
+				return p.GUID, p.Metadata, allWarnings, nil
+			}
+		}
+		return "", nil, allWarnings, actionerror.RoutePolicyNotFoundError{Source: source}
+	}
+
+	if len(routePolicies) == 0 {
+		return "", nil, allWarnings, actionerror.RoutePolicyNotFoundError{Source: ""}
+	}
+	if len(routePolicies) > 1 {
+		return "", nil, allWarnings, actionerror.RoutePolicyAmbiguityError{RouteURL: routeURL, Count: len(routePolicies)}
+	}
+	p := routePolicies[0]
+	return p.GUID, p.Metadata, allWarnings, nil
+}
 
 // RoutePolicyWithRoute combines a route policy with its associated route information
 type RoutePolicyWithRoute struct {
